fix token generation in HECEventWriter
context["token"] is set equal to the session_key (token of the splunk service).
This is not unique (two hec tokens could be created with the same splunk service and its token, making the hec not functional).

The Splunk UI (version 8+) always creates a UID as the token - the fix creates the token in the same format.