fix HECEventwriter.create_from_token
if create_from_token() is used it will call the  HECEventWriter constructor with
scheme, host and port set to None.
This will result in a call to get_splunkd_access_info() which will
throw an exception if  the hec uri does not point to the local machine.
This call is not needed in that case,
so the invocation has been moved  inside the if-statement.